requirements.txt
Django>=1.9.7
djangorestframework>=3.3.3
